+++ src/components/Select/interfaces.ts
@@ -2,10 +2,8 @@ export interface IComboProps {
   openContent: any;
   open: boolean;
   disabled?: boolean;
-  first?: boolean;
   textCombo: string;
   id: number;
-  idOpened: number | null;
 }
 
 export interface ICombo {
@@ -13,20 +11,24 @@ export interface ICombo {
   disabled?: boolean;
 }
 
-export interface IComboboxProps {
-  firstContent: any[];
-  secondContent?: any[];
-  textFirstCombo: string;
-  textSecondCombo: string;
-  textSearch: string;
+export interface ISelectProps {
+  content: any[];
+  textSelect: string;
+  searchOptions: ISearch;
   textButtonClear: string;
   textButtonConfirm: string;
   onChange: any;
   loading: boolean;
   limit: number;
   clear: boolean;
-  firstItemRender: (item: any) => void;
-  secondItemRender: (item: any) => void;
+  multiple: boolean;
+  ItemRender: (item: any) => void;
+}
+
+export interface ISearch {
+  visible: boolean;
+  textSearch?: string;
+  field?: string;
 }
 
 export interface IContentProps {
@@ -38,12 +40,11 @@ export interface IContentProps {
   handleClear: any;
   handleConfirm: any;
   onSearch: any;
-  textSearch: string;
+  searchOptions: ISearch;
   textButtonClear: string;
   textButtonConfirm: string;
   limitReached: boolean;
-  firstContentItemRender: (item: any) => void;
-  secondContentItemRender?: (item: any) => void;
+  itemRender: (item: any) => void;
 }
 
 export interface IItemMultipleProps {
@@ -72,13 +73,7 @@ export interface IInitialValues {
   open: boolean;
   actualContent: any[];
   selectedContent: any[];
-  selectedContent2: any[];
-  disable1: boolean;
-  disable2: boolean;
-  idOpened: number;
-  multiple: boolean;
-  content: any[];
-  content2: any[];
+  disable: boolean;
   disableButtonsContent: boolean;
   isLoading: boolean;
   limitReached: boolean;
@@ -89,16 +84,18 @@ export interface IReducer {
 }
 
 export enum ActionType {
-  SET_ISLOADING = "SET_ISLOADING",
-  SET_FIRST_CONTENT = "SET_FIRST_CONTENT",
-  SET_SECOND_CONTENT = "SET_SECOND_CONTENT",
-  SET_SECOND_CONTENT_SELECTED = "SET_SECOND_CONTENT_SELECTED",
-  SET_CLEAR = "SET_CLEAR",
-  OPEN_CONTENT = "OPEN_CONTENT",
-  SELECT_ITEM = "SELECT_ITEM",
-  CLEAR_FIRST_SELECTED = "CLEAR_FIRST_SELECTED",
-  CLEAR_SECOND_SELECTED = "CLEAR_SECOND_SELECTED",
-  CONFIRM_SELECT = "CONFIRM_SELECT",
-  SET_ACTUAL_CONTENT = "SET_ACTUAL_CONTENT",
-  SET_OPEN = "SET_OPEN",
+    SET_SINGLE_ITEM = "SET_SINGLE_ITEM",
+    SET_MULTIPLE_ITEM = "SET_MULTIPLE_ITEM",
+    SET_ISLOADING = "SET_ISLOADING",
+    SET_FIRST_CONTENT = "SET_FIRST_CONTENT",
+    SET_SECOND_CONTENT = "SET_SECOND_CONTENT",
+    SET_SECOND_CONTENT_SELECTED = "SET_SECOND_CONTENT_SELECTED",
+    SET_CLEAR = "SET_CLEAR",
+    OPEN_CONTENT = "OPEN_CONTENT",
+    SELECT_ITEM = "SELECT_ITEM",
+    CLEAR_FIRST_SELECTED = "CLEAR_FIRST_SELECTED",
+    CLEAR_SECOND_SELECTED = "CLEAR_SECOND_SELECTED",
+    CONFIRM_SELECT = "CONFIRM_SELECT",
+    SET_ACTUAL_CONTENT = "SET_ACTUAL_CONTENT",
+    SET_OPEN = "SET_OPEN",
 }